+++ src/Helpers.cc
@@ -43,6 +43,22 @@ namespace ignition
       }
       return false;
     }
+
+    //////////////////////////////////////////////////
+    std::vector<std::string> split(const std::string &_orig, char _delim)
+    {
+      std::vector<std::string> pieces;
+      size_t pos1 = 0;
+      size_t pos2 = _orig.find(_delim);
+      while (pos2 != std::string::npos)
+      {
+        pieces.push_back(_orig.substr(pos1, pos2-pos1));
+        pos1 = pos2 + 1;
+        pos2 = _orig.find(_delim, pos2 + 1);
+      }
+      pieces.push_back(_orig.substr(pos1, _orig.size()-pos1));
+      return pieces;
+    }
     }
   }
 }